AF-258 rename annotations to annotation.
