+++ src/main/java/com/aline/core/dto/request/OtpAuthentication.java
@@ -1,6 +1,6 @@
 package com.aline.core.dto.request;
 
-import com.aline.core.validation.annotations.Username;
+import com.aline.core.validation.annotation.Username;
 import lombok.Builder;
 import lombok.Data;
