+++ src/main/java/com/aline/core/dto/request/ResetPasswordAuthentication.java
@@ -1,7 +1,7 @@
 package com.aline.core.dto.request;
 
 import com.aline.core.dto.response.ContactMethod;
-import com.aline.core.validation.annotations.Username;
+import com.aline.core.validation.annotation.Username;
 import lombok.Builder;
 import lombok.Data;
