+++ src/main/java/com/aline/core/dto/request/ResetPasswordRequest.java
@@ -1,7 +1,7 @@
 package com.aline.core.dto.request;
 
-import com.aline.core.validation.annotations.Password;
-import com.aline.core.validation.annotations.Username;
+import com.aline.core.validation.annotation.Password;
+import com.aline.core.validation.annotation.Username;
 import lombok.Builder;
 import lombok.Data;
 import org.hibernate.validator.constraints.Length;
