+++ src/main/java/com/aline/core/dto/request/UpdateApplicant.java
@@ -1,12 +1,12 @@
 package com.aline.core.dto.request;
 
 import com.aline.core.model.Gender;
-import com.aline.core.validation.annotations.Address;
-import com.aline.core.validation.annotations.DateOfBirth;
-import com.aline.core.validation.annotations.Name;
-import com.aline.core.validation.annotations.PhoneNumber;
-import com.aline.core.validation.annotations.SocialSecurity;
-import com.aline.core.validation.annotations.Zipcode;
+import com.aline.core.validation.annotation.Address;
+import com.aline.core.validation.annotation.DateOfBirth;
+import com.aline.core.validation.annotation.Name;
+import com.aline.core.validation.annotation.PhoneNumber;
+import com.aline.core.validation.annotation.SocialSecurity;
+import com.aline.core.validation.annotation.Zipcode;
 import lombok.Builder;
 import lombok.Data;
 import org.springframework.lang.Nullable;
