+++ src/main/java/com/aline/core/validation/annotation/Name.java
@@ -1,6 +1,6 @@
-package com.aline.core.validation.annotations;
+package com.aline.core.validation.annotation;
 
-import com.aline.core.validation.validators.ZipcodeValidator;
+import com.aline.core.validation.validators.NameValidator;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
@@ -11,16 +11,18 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * String must be a ZIP code in 5 digit or +4 format.
+ * <p>The string has to be a name that contains only alphabetic letters, while also allowing hyphens and spaces.</p>
  */
 @Target({ ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = ZipcodeValidator.class)
+@Constraint(validatedBy = NameValidator.class)
 @Documented
-public @interface Zipcode {
-    String message() default "'${validatedValue}' is not a valid zipcode.";
+public @interface Name {
+
+    String message() default "Name may only contain alphabetic letters as well as hyphens and spaces.";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
+
 }
