+++ src/main/java/com/aline/core/validation/annotation/Username.java
@@ -1,6 +1,6 @@
-package com.aline.core.validation.annotations;
+package com.aline.core.validation.annotation;
 
-import com.aline.core.validation.validators.PasswordValidator;
+import com.aline.core.validation.validators.UsernameValidator;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
@@ -12,11 +12,11 @@ import java.lang.annotation.Target;
 
 @Target({ ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE })
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = PasswordValidator.class)
+@Constraint(validatedBy = UsernameValidator.class)
 @Documented
-public @interface Password {
+public @interface Username {
 
-    String message() default "Password must be at least 8 characters long and contain at least one uppercase letter, one lowercase letter, one number, and one special character.";
+    String message() default "Username must be between 6-20 characters and only include letters, numbers, underscores, and periods.";
 
     Class<?>[] groups() default {};
