+++ src/main/java/com/aline/core/validation/annotation/Zipcode.java
@@ -1,6 +1,6 @@
-package com.aline.core.validation.annotations;
+package com.aline.core.validation.annotation;
 
-import com.aline.core.validation.validators.SocialSecurityValidator;
+import com.aline.core.validation.validators.ZipcodeValidator;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
@@ -11,18 +11,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * String must be a well-formed Social Security number.
+ * String must be a ZIP code in 5 digit or +4 format.
  */
 @Target({ ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = SocialSecurityValidator.class)
+@Constraint(validatedBy = ZipcodeValidator.class)
 @Documented
-public @interface SocialSecurity {
-
-    String message() default "${validatedValue} is not a valid Social Security number.";
+public @interface Zipcode {
+    String message() default "'${validatedValue}' is not a valid zipcode.";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
-
 }
