+++ src/main/java/com/aline/core/validation/validators/DateOfBirthValidator.java
@@ -1,6 +1,6 @@
 package com.aline.core.validation.validators;
 
-import com.aline.core.validation.annotations.DateOfBirth;
+import com.aline.core.validation.annotation.DateOfBirth;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
