+++ src/main/java/com/aline/core/validation/validators/MembershipIdValidator.java
@@ -1,7 +1,7 @@
 package com.aline.core.validation.validators;
 
 
-import com.aline.core.validation.annotations.MembershipId;
+import com.aline.core.validation.annotation.MembershipId;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
