+++ src/main/java/com/aline/core/validation/validators/NameValidator.java
@@ -1,6 +1,6 @@
 package com.aline.core.validation.validators;
 
-import com.aline.core.validation.annotations.Name;
+import com.aline.core.validation.annotation.Name;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
