+++ src/main/java/com/aline/core/validation/validators/SocialSecurityValidator.java
@@ -1,6 +1,6 @@
 package com.aline.core.validation.validators;
 
-import com.aline.core.validation.annotations.SocialSecurity;
+import com.aline.core.validation.annotation.SocialSecurity;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
