+++ src/main/java/com/aline/core/validation/validators/ZipcodeValidator.java
@@ -1,6 +1,6 @@
 package com.aline.core.validation.validators;
 
-import com.aline.core.validation.annotations.Zipcode;
+import com.aline.core.validation.annotation.Zipcode;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
